requirements.txt
selenium==4.15.0
pytest==7.4.3
webdriver-manager==4.0.1
allure-pytest==2.13.2
python-dotenv==1.0.0
requests==2.31.0
pytest-xdist==3.5.0

